requirements.txt
PyYAML
slugify
requests
jinja2
humanize
beautifulsoup4
